DAOS-16627 test: Automate change fabric provider tests
Test the changing fabric provider without reformatting storage feature.
The first test writes to a file using IOR, changes the fabric provider
and reads/checks the file for correctness.
The second test does the same tests and then reverts the system provider
back to the original provider and tries to read the file.

Quick-functional: true
