+++ src/tests/ftest/control/change_fabric_provider.yaml
@@ -1,6 +1,6 @@
 hosts:
   test_servers: 3
-  test_clients: 1
+  test_clients: 2
 setup:
   start_agents_once: false
   start_servers_once: false
@@ -15,10 +15,8 @@ server_config:
   system_ram_reserved: 2
 pool:
   size: 90%
-  svcn: 1
 container:
   type: POSIX
-  control_method: daos
 ior:
   env_vars:
     - D_LOG_MASK=INFO
@@ -26,13 +24,10 @@ ior:
     - DD_SUBSYS=all
   api: POSIX
   client_processes:
-    np_16:
-      np: 16
+    np: 16
   test_file: testFile
-  repetitions: 1
   read_flags: "-v -C -k -e -r -R -G 27"
-  iorflags:
-    flags: "-v -k -e -w -r -R -G 27"
+  flags: "-v -k -e -w -r -R -G 27"
   transfer_size: '1M'
   block_size: '100M'
   write_x: 2
